+++ src/main/java/io/zipcoder/microlabs/mastering_loops/Numbers.java
@@ -1,101 +1,64 @@
 package io.zipcoder.microlabs.mastering_loops;
 
-import org.junit.Assert;
-import org.junit.Test;
 
-public class NumbersTest {
+public class Numbers {
 
-    @Test
-    public void oneToTenTest(){
-        //: Given
-        Numbers numbers = new Numbers();
-        String expected = "*** Output ***\noneToTen()\n1\n2\n3\n4\n5\n6\n7\n8\n9\n10";
+    public String oneToTen(){
+        String output = "*** Output ***\noneToTen()";
+        for(int i = 1; i< 11; i++){
+            output += "\n"+ i;
+        }
 
-        //: When
-        String actual = numbers.oneToTen();
-
-        //: Then
-
-        // System.out.println("expected\n" + expected);
-        // System.out.println("actual\n" + actual);
-        Assert.assertEquals("The two strings are equal", expected, actual);
-    }
-
-    @Test
-    public void oddNumbersTest(){
-
-        //: Given
-        Numbers numbers = new Numbers();
-        String expected = "oddNumbers()\n*** Output ***\n1\n3\n5\n7\n9\n11\n13\n15\n17\n19";
-
-        //: When
-        String actual = numbers.oddNumbers();
-
-        //: Then
-        Assert.assertEquals("Strings are equal ", expected, actual);
-    }
-
-    @Test
-    public void squaresTest(){
-        //: Given
-        Numbers numbers = new Numbers();
-        String expected = "squares()\n*** Output ***\n1\n4\n9\n16\n25\n36\n49\n64\n81\n100";
-
-        //: When
-        String actual = numbers.squares();
-
-
-        //: Then
-        Assert.assertEquals("Strings are equal ", expected, actual);
-
+        return output;
 
     }
 
-    @Test
-    public void random4Test(){
-        //: Given
-        Numbers numbers = new Numbers();
-        String expected = "random4()\n*** Output ***\n1\n4\n9\n16";
+    public String oddNumbers(){
+        String output = "oddNumbers()\n*** Output ***";
+        for(int i = 1; i<21; i++){
+            if(i%2 != 0){
+                output += "\n" + i;
+            }
 
-        //: When
-        String actual = numbers.random4();
-
-
-        //: Then
-        Assert.assertEquals("Strings are equal ", expected, actual);
-
+        }
+        return output;
     }
 
-    @Test
-    public void evenTest(){
-        //: Given
-        Numbers numbers = new Numbers();
-
-        String expected = "even(20)\n*** Output ***\n2\n4\n6\n8\n10\n12\n14\n16\n18";
-
-        //: When
-        String actual = numbers.even(20);
+    public String squares(){
+        String output = "squares()\n*** Output ***";
+        for(int i = 1; i<11; i++){
+            output += "\n" + i*i;
 
 
-        //: Then
-        Assert.assertEquals("Strings are equal ", expected, actual);
-
+        }
+        return output;
     }
 
-    @Test
-    public void powersTest(){
-        //: Given
-        Numbers numbers = new Numbers();
-        String expected = "powers(8)\n*** Output ***\n2\n4\n8\n16\n32\n64\n128\n256";
-
-        //: When
-        String actual = numbers.powers(8);
+    public String random4(){
+        String output = "random4()\n*** Output ***";
+        for(int i = 0; i<4; i++){
+            output += "\n" + (int)((Math.random()*9)+1);
 
 
-        //: Then
-        Assert.assertEquals("Strings are equal ", expected, actual);
-
-
+        }
+        return output;
     }
 
+    public String even(int n){
+        String output = "even(" + n + ")\n*** Output ***";
+        for(int i = 1; i<n; i++){
+            if(i%2 == 0) {
+                output += "\n" + i;
+            }
+        }
+        return output;
+    }
+
+    public String powers(int n) {
+        String output = "powers("+n+")\n*** Output ***";
+        for (int i = 1; i <= n; i++) {
+            output += "\n" + (int) Math.pow(2, i);
+        }
+        return output;
+    }
 }
